fix ShortBuffer tests
The error handling for short buffers is looking for a panic on indexing. The actuall string value has changed in Go versions and broke this handling.  This fixes it short term by looking for a substring of the error instead of exactly matcing.  Long term this panic/recovery needs to be fixed and we should know the size of the packet being read.